requirements.txt
# Core
pandas>=2.0
requests>=2.30
streamlit>=1.37
plotly>=5.0
streamlit-aggrid>=1.0

# ML
scikit-learn>=1.3

# Database
sqlalchemy>=2.0
pyodbc>=5.0

# Azure
azure-identity>=1.15
azure-keyvault-secrets>=4.7

# Configuration
python-dotenv>=1.0

# Development (optional)
matplotlib>=3.8
seaborn>=0.13
jupyterlab>=4.0
